requirements.txt
python-constraint
svgwrite
CairoSVG
Pillow
reportlab